requirements.txt
torch>=1.9
torchvision>=0.10
numpy
scikit-learn
imbalanced-learn
PyYAML
matplotlib
scipy
PyWavelets
Pillow
opencv-python-headless
wandb
